feat: [website-test-breakdown-004] - [Extract layout, graph, submit-work, and trace follow-up coverage until App.test.tsx fits standard limits]
